LayoutTests/fast/dom/Geolocation/reentrant-permission-denied-expected.txt

Tests that reentrant calls to Geolocation methods from the error callback due to a PERMISSION_DENIED error are OK.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS error.code is error.PERMISSION_DENIED
PASS error.message is "User denied Geolocation"
PASS error.code is error.PERMISSION_DENIED
PASS error.message is "User denied Geolocation"
PASS successfullyParsed is true

TEST COMPLETE
